refactor: tech-stacks 디렉토리 projects 안으로 변경
