+++ src/app.module.ts
@@ -23,7 +23,6 @@ import { UsersModule } from './users/users.module';
 import { AuthenticationModule } from './authentication/authentication.module';
 import { ProjectsModule } from './projects/projects.module';
 import { ApplicationsModule } from './applications/applications.module';
-import { ProjectsTechStacksModule } from './tech-stacks/projects-tech-stacks.module';
 
 const ENV = process.env;
 const entities = [
@@ -61,7 +60,6 @@ const entities = [
     AuthenticationModule,
     UsersModule,
     ProjectsModule,
-    ProjectsTechStacksModule,
     ApplicationsModule,
   ],
   providers: [
