+++ src/projects/projects.module.ts
@@ -5,9 +5,14 @@ import { ProjectsLikeModule } from './like/projects-like.module';
 import { ProjectsController } from './projects.controller';
 import { ProjectsService } from './projects.service';
 import { ProjectsRepository } from './repository';
+import { ProjectsTechStacksModule } from './tech-stacks/projects-tech-stacks.module';
 
 @Module({
-  imports: [ProjectsMembersModule, ProjectsLikeModule],
+  imports: [
+    ProjectsMembersModule,
+    ProjectsLikeModule,
+    ProjectsTechStacksModule,
+  ],
   controllers: [ProjectsController],
   providers: [ProjectsService, ProjectsRepository],
   exports: [ProjectsService],
