+++ src/projects/tech-stacks/dto/delete-tech-stacks-request.dto.ts
