+++ src/projects/tech-stacks/projects-tech-stacks.module.ts
@@ -3,10 +3,8 @@ import { Module } from '@nestjs/common';
 import { ProjectsTechStacksController } from './projects-tech-stacks.controller';
 import { ProjectsTechStacksRepository } from './repository';
 import { ProjectsTechStacksService } from './projects-tech-stacks.service';
-import { ProjectsModule } from '../projects/projects.module';
 
 @Module({
-  imports: [ProjectsModule],
   controllers: [ProjectsTechStacksController],
   providers: [ProjectsTechStacksService, ProjectsTechStacksRepository],
   exports: [ProjectsTechStacksService, ProjectsTechStacksRepository],
